Fix docker builds version/builder, new targets
The Docker build was not using the BUILDER and BUILD_VERSION
env vars from the host so those fields in the version info
was missing.

Added fields to build the application in Docker with no caching,
a clean environment, and added build caching to the quick build.

Added target to rebuild, restart, and show logs.