Implement entrees with sides
Now, when a new item is added to an order, if it has a
non-zero number of sides, the UI will redirect to the sides
category and the selected items will be added to the OrderItem.
The current order component has been updated to display entrees
with their sides distinguishing them by font weight and indention.
The price of sides is not take into account when they are part
of an entree.